refactor: unify package task
