+++ pkg/domain/task/remind_time_test.go
@@ -1,4 +1,4 @@
-package remindtime_test
+package task_test
 
 import (
 	"errors"
